Add POM generation for meta-plugins
This commit adds a POM generation task for the ZIP distribution for
meta-plugins. With this change we also modify MetaPluginBuildPlugin to
extend BuildPlugin (because we need POM generation configuration from
there), remove standalone and REST test plugins, and manually create an
integTest task. To simplify this, we abstract out some shared code from
PluginBuildPlugin so that it can be reused in MetaPluginBuildPlugin.
